tests(cli): reorganize by feature so read-only stays read-only
